Add showing images in container
- upload, show images with new templates
