+++ client/package.json
@@ -13,6 +13,8 @@
     "antd": "^4.16.13",
     "axios": "^0.24.0",
     "fs": "^0.0.1-security",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-minimal-side-navigation": "^1.9.2",
