--- client/src/album.js
@@ -1,121 +1,0 @@
-import * as React from 'react';
-import AppBar from '@mui/material/AppBar';
-import Button from '@mui/material/Button';
-import Card from '@mui/material/Card';
-import CardActions from '@mui/material/CardActions';
-import CardContent from '@mui/material/CardContent';
-import CardMedia from '@mui/material/CardMedia';
-import CssBaseline from '@mui/material/CssBaseline';
-import Grid from '@mui/material/Grid';
-import Stack from '@mui/material/Stack';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import Container from '@mui/material/Container';
-import Link from '@mui/material/Link';
-import SearchIcon from '@mui/icons-material/Search';
-import { styled, alpha } from '@mui/material/styles';
-import InputBase from '@mui/material/InputBase';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-
-import logo from './logo.png';
-import './album.css';
-import { Search, SearchIconWrapper, StyledInputBase, PrimarySearchAppBar } from './component/SearchBar';
-
-
-const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
-
-const theme = createTheme();
-
-export default function Album() {
-  return (
-    <ThemeProvider theme={theme}>
-      {/* <CssBaseline /> */}
-      {/* <AppBar position="relative">
-        <Toolbar className="title">
-          <img src={logo} className="icon" alt="logo" sx={{ mr: 5 }} />
-          <Typography variant="h"color="inherit" noWrap>
-            NOONBODY
-          </Typography>
-          <Search >
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Search…"
-              inputProps={{ 'aria-label': 'search' }}
-            />
-            </Search>
-        </Toolbar>
-      </AppBar> */}
-      <PrimarySearchAppBar />
-      
-      <main>
-        {/* Hero unit */}
-        {/* <Box
-          sx={{
-            bgcolor: 'background.paper',
-            pt: 8,
-            pb: 6,
-          }}
-        >
-          <Container maxWidth="sm">
-            <Stack
-              sx={{ pt: 4 }}
-              direction="row"
-              spacing={2}
-              justifyContent="center"
-            >
-              <Button variant="contained">Main call to action</Button>
-              <Button variant="outlined">Secondary action</Button>
-            </Stack>
-          </Container>
-        </Box> */}
-        <Container sx={{ py: 8 }} maxWidth="md">
-          {Date.now()}
-          <Grid container spacing={4}>
-            {cards.map((card) => (
-              <Grid item key={card} xs={3} sm={3} md={3}>
-                <Card
-                  sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}
-                >
-                  <CardMedia
-                    component="img"
-                    sx={{
-                      // 16:9
-                    //   pt: '56.25%',
-                    }}
-                    image="https://source.unsplash.com/random"
-                    alt="random"
-                  />
-                  <CardContent sx={{ flexGrow: 1 }}>
-                    {/* <Typography>
-                      This is a media card. You can use this section to describe the
-                      content.
-                    </Typography> */}
-                  </CardContent>
-                  <CardActions>
-                    <Button size="small">DELETE</Button>
-                    <Button size="small">Edit</Button>
-                  </CardActions>
-                </Card>
-              </Grid>
-            ))}
-          </Grid>
-        </Container>
-      </main>
-      {/* Footer */}
-      <Box sx={{ bgcolor: 'background.paper', p: 6 }} component="footer">
-        <Typography
-          variant="subtitle1"
-          align="center"
-          color="text.secondary"
-          component="p"
-        >
-          Images from Instagram @cbum.
-        </Typography>
-      </Box>
-      {/* End footer */}
-    </ThemeProvider>
-  );
-}
