+++ client/src/component/UploadImage.css
@@ -6,5 +6,5 @@
 
 img {
     height: 200px;
-    width: 200px;
+    width: auto;
   }
